Refactor: Compartimentalize migration helpers
