+++ specifyweb/specify/migration_utils/migration_helpers/0018_cot_catnum_schema_helper.py
@@ -1,0 +1,60 @@
+from specifyweb.specify.migration_utils.schema_reader import camel_to_spaced_title_case
+from specifyweb.specify.models import datamodel
+
+# ##########################################
+# Used in 0018_cot_catnum_schema.py
+# ##########################################
+
+def add_cot_catnum_to_schema(apps, schema_editor=None):
+    Splocalecontainer = apps.get_model('specify', 'Splocalecontainer')
+    Splocalecontaineritem = apps.get_model('specify', 'Splocalecontaineritem')
+    Splocaleitemstr = apps.get_model('specify', 'Splocaleitemstr')
+
+    CollectionObjectType_Table = datamodel.get_table_strict(
+        'collectionobjecttype')
+    catalognumber_format_field = CollectionObjectType_Table.get_field_strict(
+        'catalogNumberFormatName')
+
+    for container in Splocalecontainer.objects.filter(name='collectionobjecttype', schematype=0):
+        schema_item, created = Splocalecontaineritem.objects.get_or_create(
+            name=catalognumber_format_field.name, type=catalognumber_format_field.type, container=container)
+        if created:
+            schema_item.version = 0
+
+        schema_item.isrequired = (
+            catalognumber_format_field.required
+            if schema_item.isrequired is None
+            else schema_item.isrequired
+        )
+
+        schema_item.save()
+
+        schema_name = camel_to_spaced_title_case(
+            catalognumber_format_field.name)
+        Splocaleitemstr.objects.get_or_create(
+            language='en', text=schema_name, itemname=schema_item)
+        Splocaleitemstr.objects.get_or_create(
+            language='en', text=schema_name, itemdesc=schema_item)
+
+def remove_cot_catnum_from_schema(apps, schema_editor=None):
+    Splocalecontainer = apps.get_model('specify', 'Splocalecontainer')
+    Splocalecontaineritem = apps.get_model('specify', 'Splocalecontaineritem')
+    Splocaleitemstr = apps.get_model('specify', 'Splocaleitemstr')
+
+    CollectionObjectType_Table = datamodel.get_table_strict(
+        'collectionobjecttype')
+    catalognumber_format_field = CollectionObjectType_Table.get_field_strict(
+        'catalogNumberFormatName')
+
+    containers = Splocalecontainer.objects.filter(
+        name='collectionobjecttype', schematype=0)
+    items = Splocalecontaineritem.objects.filter(
+        name='catalogNumberFormatName', container__in=containers)
+
+    schema_name = camel_to_spaced_title_case(catalognumber_format_field.name)
+    filters = Q(language='en', text=schema_name) & (
+        Q(itemname__in=items) | Q(itemdesc__in=items))
+    locale_strings = Splocaleitemstr.objects.filter(filters)
+
+    locale_strings.delete()
+    items.delete()
